Fix file structure from desktop app push
