+++ control/templates/main_page.html
@@ -8,6 +8,7 @@
     <h1>
       PrettyPi
       <small>Version 0.0.1</small>
+      <h5>Click "Tasks" to view ongoing tasks</h5>
     </h1>
     <ol class="breadcrumb">
       <li><a href="#"><i class="fa fa-dashboard"></i> Home</a></li>
