--- main_page.html
@@ -1,26 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<!-- Content Wrapper. Contains page content -->
-<div class="content-wrapper">
-  <!-- Content Header (Page header) -->
-  <section class="content-header">
-    <h1>
-      PrettyPi
-      <small>Version 0.0.1</small>
-      <h5>Click "Tasks" to view ongoing tasks</h5>
-    </h1>
-    <ol class="breadcrumb">
-      <li><a href="#"><i class="fa fa-dashboard"></i> Home</a></li>
-      <li class="active">Dashboard</li>
-    </ol>
-  </section>
-
-  <!-- Main content -->
-  <section class="content">
-  </section>
-  <!-- /.content -->
-</div>
-<!-- /.content-wrapper -->
-
-{% endblock %}
